script to upgrade from Peggy 1 to Peggy 2
